requirements.txt
asgiref==3.7.2
boto3==1.33.9
botocore==1.33.9
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.3.2
cryptography==41.0.7
dj-database-url==2.1.0
Django==5.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-heroku==0.3.1
django-storages==1.14.2
djangorestframework==3.14.0
environ==1.0
gunicorn==21.2.0
idna==3.6
iniconfig==2.0.0
jmespath==1.0.1
packaging==23.2
pillow==10.2.0
pluggy==1.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pytest==7.4.3
pytest-django==4.7.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
responses==0.24.1
s3transfer==0.8.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7
whitenoise==6.6.0
